+++ test/unit/reactive-document.test.ts
@@ -20,10 +20,10 @@ import {
     first
 } from 'rxjs/operators';
 import type {
-    RxChangeEvent,
-    RxDocument
+    RxChangeEvent
 } from '../../src/types';
 import { HumanDocumentType } from '../helper/schemas';
+import { firstValueFrom } from 'rxjs';
 
 config.parallel('reactive-document.test.js', () => {
     describe('.save()', () => {
@@ -143,6 +143,19 @@ config.parallel('reactive-document.test.js', () => {
         });
         describe('negative', () => { });
     });
+    describe('.$', () => {
+        it('should emit a RxDocument, not only the document data', async () => {
+            const c = await humansCollection.create(1);
+            const doc = await c.findOne().exec(true);
+
+            const firstEmitPromise = firstValueFrom(doc.$);
+            doc.incrementalPatch({ age: 100 });
+
+            const emitted = await firstEmitPromise;
+            assert.ok(emitted.$);
+            c.database.destroy();
+        });
+    });
     describe('.get$()', () => {
         describe('positive', () => {
 
@@ -182,47 +195,5 @@ config.parallel('reactive-document.test.js', () => {
         });
     });
     describe('issues', () => {
-        it('#3434 event data must not be mutateable', async () => {
-            const db = await createRxDatabase({
-                name: randomCouchString(10),
-                storage: config.storage.getStorage(),
-                eventReduce: true,
-                ignoreDuplicate: true
-            });
-            const collections = await db.addCollections({
-                mycollection: {
-                    schema: schemas.humanDefault
-                }
-            });
-            await collections.mycollection.insert({
-                passportId: 'foobar',
-                firstName: 'Bob',
-                lastName: 'Kelso',
-                age: 56
-            });
-
-            const person: RxDocument<HumanDocumentType> = await db.mycollection.findOne().exec();
-
-
-            let hasThrown = false;
-            person.$.subscribe(data => {
-                try {
-                    // mutating the document data is not allowed and should throw
-                    delete (data as any)['_rev'];
-                } catch (err) {
-                    hasThrown = true;
-                }
-            });
-
-            await person.incrementalModify(state => {
-                state.age = 50;
-                return state;
-            });
-
-            assert.strictEqual(person.getLatest().age, 50);
-            assert.ok(hasThrown);
-
-            db.destroy();
-        });
     });
 });